+++ src/FairPlayCombinedSln/FairPlayCombined.Models/FairPlayDating/Religion/ReligionModel.cs
@@ -10,7 +10,7 @@ namespace FairPlayCombined.Models.FairPlayDating.Religion
 {
     public class ReligionModel : IListModel
     {
-        public short? ReligionId { get; set; }
+        public int? ReligionId { get; set; }
         public string? Name { get; set; }
     }
 }
