+++ src/FairPlayCombinedSln/FairPlayCombinedDb/FairPlayDating/TattooStatus.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [FairPlayDating].[TattooStatus]
 (
-	[TattooStatusId] SMALLINT NOT NULL CONSTRAINT PK_TattooStatus PRIMARY KEY IDENTITY, 
+	[TattooStatusId] INT NOT NULL CONSTRAINT PK_TattooStatus PRIMARY KEY IDENTITY, 
     [Name] NVARCHAR(50) NOT NULL
 )
